Toleration for Java 9 and 10
As of Java 9 the $JAVA_HOME/lib/tools.jar was removed and converted
to a module. Made the dependency on the jar option for JDK 9+. Also,
upgrading the default compiler compliance to JDK 1.7 since javac no
longer supports -source 1.5 as of JDK 9.  Liberty itself requires a
minimum of JDK 1.7 as well.

Signed-off-by: Andrew Guibert <andy.guibert@gmail.com>