Update WLPManagedContainer.java
Pad the application name with spaces so that when it searches for an application name, it doesn't find one that is really a different app name.  For example, I got the "this should never happen" message because it searched for `CWWKZ0001I.*ExceptionMapperTest` - and found this line: `CWWKZ0001I: Application DefaultExceptionMapperTest started in 2.086 seconds.`

This change should prevent this type of mismatch.