Revert "Update WLPManagedContainer.java"
This reverts commit 93c164b8fe2dc25e1756413038ed233601f62569.

This was an attempt to find the correct log message where two apps were
deployed and the name of one was a substring of the other.

Unfortunately, in some log messages, the application name is not
surrounded by spaces so this change resulted in the correct log line not
being found at all in some cases.