Fix remote container tests for 19.0.0.3
The test is currently set up to expect the server to generate a key.jks
file. Since 19.0.0.3, liberty defaults to generating a key.p12 file
instead.

As a quick fix, change the server config to specify a key.jks file as
the test expects.

We can't change the server config to always create key.p12 as earlier
versions of liberty can't generate this file format.