Changes to support Arquillian Chameleon
When Arquillian Chameleon does the Liberty install it does not preserve the file permissions
as a result when this container tries to run the server script it fails. Before starting the
check and set the execute bit.

When running server script from the ProcessBuilder the server.env file in the server config dir
does not get run. Setting the directory to the wlpHome causes it to process server.env. There is
probably a bug in Liberty here, but for now fastest, simplest to set the directory.